tidy a little
